Python/Python Bootcamp/Team_00/src/requirements.txt
beautifulsoup4==4.12.3
html2image=2.0.4.3
lxml==5.2.2
networkx==3.3
python-dotenv=1.0.1
pyvis==0.3.2
requests==2.31.0
